requirements.txt
pyinstaller
streamlit
flask
pywebview